*** empty log message ***

+++ tools/xtcdump.cc
@@ -33,12 +33,12 @@ namespace Pds {
 	return 0;
       }
       ssize_t sz = ::read(_fd, dg.xtc.payload(), payloadSize);
-      if (sz != payloadSize) {
+      if (sz != (ssize_t)payloadSize) {
 	printf("XtcFileIterator::next read incomplete payload %d/%d\n",
 	       sz,payloadSize);
       }
       
-      return sz!=payloadSize ? 0: _cdg;
+      return sz!=(ssize_t)payloadSize ? 0: _cdg;
     }
   private:
     int         _fd;
